support Frame Start/End; rework frame data model
We previously stored pre-frame and post-frame updates in separate Vecs,
one for each port. But this makes less sense with the addition of Frame
Start/End events.

Now we store frame data row-wise, with a top-level `Frame` object that
contains data for all ports for a given frame.

To keep the frame data contiguous in memory, we use const generics and
renumber ports so the lowest active port is always `ports[0]`, etc.

The `index` property of frames is also gone, as we can now reliably
translate between array indices and frame numbers via
`game::FIRST_FRAME_INDEX`.